+++ Adaptive_Systems_Project/src/Entities/Item.java
@@ -3,7 +3,7 @@ package Entities;
 import java.util.Random;
 
 public class Item extends Entity{
-	private int rating;
+	private double rating;
 	private int id;
 
 	public Item(int id) {
@@ -14,16 +14,16 @@ public class Item extends Entity{
 		rating = 0;
 		id = new Random().nextInt(1000);
 	}
-	public Item(int rating, int id) {
+	public Item(double rating, int id) {
 		this.rating = rating;
 		this.id=id;
 	}
 	
-	public int getRating() {
+	public double getRating() {
 		return rating;
 	}
 
-	public void setRating(int rating) {
+	public void setRating(double rating) {
 		this.rating = rating;
 	}
